\name{corrdat.sm}
\docType{data}
\alias{corrdat.sm}
\title{Data for Fitting Correlated Effects Model With Small-Sample Corrections}
\description{
Data used in Tipton (2013).
}
\usage{corrdat}
\format{
  A dataframe containing 300 effect sizes from 28 studies used in Tipton (2013).
}
\source{Elizabeth Tipton}
\references{
  Tipton, E. (in press) Small sample adjustments for robust variance estimation 
  with meta-regression. \emph{Psychological Methods}.
}
\keyword{datasets}