% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.center.R
\name{group.center}
\alias{group.center}
\title{Convenience function for calculating group-centered covariates.}
\usage{
group.center(var, grp)
}
\arguments{
\item{var}{The covariate to be group centered.}

\item{grp}{A vector corresponding to the group identification.}
}
\value{
A column or vector containing the group.centered covariate.
}
\description{
Creates a within-study (or within-cluster) version of the covariate in
question.
}
\examples{


# Load data
data(corrdat) 

# Create a group centered covariate 
males_c <- group.center(corrdat$males, corrdat$studyid)


}
\keyword{robumeta}
