% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.mean.R
\name{group.mean}
\alias{group.mean}
\title{Convenience function for calculating group-mean covariates.}
\usage{
group.mean(var, grp)
}
\arguments{
\item{var}{The covariate cotaining the values to be group averaged.}

\item{grp}{The group from which the average should be calculated.}
}
\value{
A column or vector containing the group.mean covariate.
}
\description{
Creates a between-study (or between-cluster) version of the covariate in
question.
}
\examples{


# Load data
data(corrdat)

# Create a group mean covariate 
age_m <- group.mean(corrdat$age, corrdat$studynum)


}
\keyword{robumeta}
