\name{hedgesdat}
\docType{data}
\alias{hedgesdat}
\title{hedgesdat}
\description{
Data from a meta-analysis on the effectiveness of phonics reading instruction by
Ehri, Nunes, Stahl and Willows (2001). Data reported in Hedges, Tipton, and 
Johnson (2010) with example.
}
\usage{hedgesdat}
\format{A dataframe containing 179 effect sizes from 66 different studies}
\source{Hedges, Tipton, and Johnson (2010)}
\references{

  Ehri, L.C., Nunes, S.R., Stahl, S.A., Willows, D.M. (2001) Systematic phonics 
  instruction helps children learn to read: Evidence from the National Reading 
  Panel's meta-analysis. \emph{Review of Educational Research}. \bold{71}, 
  393--447.

  Hedges, L.V., Tipton, E., Johnson, M.C. (2010) Robust variance estimation in 
  meta-regression with dependent effect size estimates. \emph{Research Synthesis 
  Methods}. \bold{1}(1): 39--65. Erratum in \bold{1}(2): 164--165. DOI: 
  10.1002/jrsm.5
}
  
\keyword{datasets}