% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robumeta-data.R
\docType{data}
\name{hierdat}
\alias{hierdat}
\title{Data for Fitting Hierarchical Effects Model}
\format{A dataframe containing 68 effect sizes from 15 studies used in
Tanner- Smith and Tipton (2013).}
\source{
https://my.vanderbilt.edu/emilytannersmith/training-materials/
}
\description{
Fictional data used in TannerTanner-Smith and Tipton (2013).
}
\references{
Tanner-Smith E.E., Tipton, E. (2013) Robust variance estimation
with dependent effect sizes: practical considerations including a software
tutorial in Stata and SPSS. \emph{Research Synthesis Methods}. ISSN
1759-2887.
}
\keyword{datasets}
