% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.robu.R
\name{predict.robu}
\alias{predict.robu}
\title{Predict method for robumeta model fits.}
\usage{
\method{predict}{robu}(object, pred_vector, level = 0.95)
}
\arguments{
\item{object}{A fitted robumeta model object.}

\item{pred_vector}{A prediction vector for hypothesis test. This contains the covariate values we have. The conditional mean value, as well as its confidence interval, of the outcome (dependent) variable will be predicted given the values.

 INTERCEPT: If intercept is included in the robumeta model, the first element should always be 1, which represents the intercept, and followed by the covariate values in order. If the robumeta model does non have an intercept, the prediction vector should begin with the first covariate value.

 VARIABLE TYPE: For a continuous variable, we just use the variable value as the corresponding element value in the vector. For a categorical variable, the original variable value should be transformed by the coding system identical to the one used in the robumeta model.(Dummy coding, deviation coding, etc.)
 
 NA VALUES: If the vector contains NAs, \code{predict.robu} will remove the corresponding covariates from the original data, and refit a new robumeta model.The prediction and confidence interval will be estimated based on that new model.}

\item{level}{Tolerance/confidence level.}
}
\value{
\code{predict.robu}   produces a vector of prediction and confidence interval with element names \code{prediction, se, df, t_score, lowerBound,} and \code{upperBound}.

\code{prediction:}   the predicted value based on the prediction vector.

\code{se:}  standard error of predicted mean

\code{t_score:} The t-statistic calculated based on the predicted mean.

\code{df:}  The small sample corrected degrees of freedom of the distribution of the t-statistic.

\code{lowerBound:} lower bound of the confidence interval for the predicted mean.

\code{upperBound:} upper bound of the confidence interval for the predicted mean.
}
\description{
\code{predict.robu} obtains the predicted outcome mean and confidence interval of a fitted robumeta model object given a prediction vector.
}
\examples{
require(robumeta)
require(clubSandwich)

## Fit a robumeta model by robu() function. 



robu_mod <- robu(LOR1 ~ study_design + duration + service_hrs, 
                 data = dropoutPrevention, studynum = studyID, var.eff.size = varLOR, 
                 modelweights = "HIER",small = T)

## Construct the prediction vector for the predicted mean given 
## study_design = "Matched", duration = 38 and service_rs = 5.
##
## In this robumeta model, the first covariate is a categorical variable that contains three levels:
## "Matched" (33\%, dummy code: 00), "Randomized"(24\%, 01) and "non-match non-randomized"(43\%, 10).
##
## The corresponding prediction vector begins with 1 (intercept), and followed by 0 0, the dummy code for "Mateched".
## The last two elements are 38 and 5, which are the values for duration and sevice_hrs.
                                           
predict(object = robu_mod,pred_vector = c(1,0,0,38,5),level = 0.95)

## If we do not know the value of duration, the prediction vector should be c(1,0,0,NA,5). 
## predict.robu() will refit a new model without the covariate duration, and the prediction will be based on it.

predict(object = robu_mod,pred_vector = c(1,0,0,NA,5),level = 0.95)

}

