\name{robumeta}
\alias{robumeta-package}
\docType{package}
\title{robumeta}

\description{\pkg{robumeta} is used to fit meta-regression models using robust 
  variance estimation (RVE) methods for correlated and hierarchical effects 
  models using the original (Hedges, Tipton and Johnson, 2010) and small-sample 
  corrected (Tipton, 2013) methods. 
}
\details{
\tabular{ll}{
Package: \tab robumeta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-17\cr
License: \tab GPL-2\cr
}
}

\author{
  Zachary Fisher and Elizabeth Tipton
}



\references{

  Hedges, L.V., Tipton, E., Johnson, M.C. (2010) Robust variance estimation in 
  meta-regression with dependent effect size estimates. \emph{Research Synthesis 
  Methods}. \bold{1}(1): 39--65. Erratum in \bold{1}(2): 164--165. DOI: 
  10.1002/jrsm.5

  Tipton, E. (in press) Small sample adjustments for robust variance estimation 
  with meta-regression. \emph{Psychological Methods}.
}


\keyword{robu}