% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robumeta-data.R
\docType{data}
\name{corrdat.sm}
\alias{corrdat.sm}
\title{Data for Fitting Correlated Effects Model With Small-Sample Corrections}
\format{A dataframe containing 300 effect sizes from 28 studies used in
Tipton (2013).}
\source{
Elizabeth Tipton
}
\description{
Data used in Tipton (2013).
}
\references{
Tipton, E. (in press) Small sample adjustments for robust
variance estimation with meta-regression. \emph{Psychological Methods}.
}
\keyword{datasets}
