\name{hierdat}
\docType{data}
\alias{hierdat}
\title{Data for Fitting Hierarchical Effects Model}
\description{
Fictional data used in TannerTanner-Smith and Tipton (2013).
}
\usage{hierdat}
\format{
  A dataframe containing 68 effect sizes from 15 studies used in Tanner-
  Smith and Tipton (2013).
}
\source{https://my.vanderbilt.edu/emilytannersmith/training-materials/}
\references{
  Tanner-Smith E.E., Tipton, E. (2013) Robust variance estimation with dependent 
  effect sizes: practical considerations including a software tutorial in Stata 
  and SPSS. \emph{Research Synthesis Methods}. ISSN 1759-2887.
}
\keyword{datasets}