% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.robu.R
\name{predict.robu}
\alias{predict.robu}
\title{Predict method for robumeta model fits.}
\usage{
\method{predict}{robu}(object, contrast_vector, level = 0.95)
}
\arguments{
\item{object}{A fitted robumeta model object.}

\item{level}{Tolerance/confidence level.}

\item{constraints_vector}{A contrast vector to test. It the vector contains NAs, \code{predict.robu} will remove the corresponding covariates from the original data, and refit a new robumeta model.The prediction and prediction interval will be estimated based on that new model.}

\item{\code{prediction}}{the predicted value}

\item{\code{se}}{standard error of predicted mean}

\item{\code{t_score}}{The t-statistic calculated based on the predicted mean.}

\item{\code{df}}{The small sample corrected degrees of freedom of the distribution of the t-statistic.}

\item{\code{lowerBound}}{lower bound of the confidence interval for the predicted mean.}

\item{\code{upperBound}}{upper bound of the confidence interval for the predicted mean.}
}
\value{
\code{predict.robu }   produces a vector of prediction and confidence interval with element names \code{prediction, se, df, t_score, lowerBound,} and \code{upperBound}.
}
\description{
\code{predict.robu} obtains predictions and prediction confidence intervals of a fitted robumeta model object for a constraints vector.
}
\examples{
library(robumeta)
library(clubSandwich)

robu_mod <- robu(LOR1 ~ study_design + duration + service_hrs, 
                 data = dropoutPrevention, studynum = studyID, var.eff.size = varLOR, 
                 modelweights = "HIER",small = T)
predict(object = robu_mod,contrast_vector = c(1,1,0,0,0),level = 0.95)
}

